api/requirements.txt
google-cloud-firestore>=2.0.0
requests>=2.25.1
